requirements.txt
altgraph==0.17
anyio==2.0.2
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
fbs==0.9.0
filelock==3.0.12
future==0.18.2
glcontext==2.2.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==2.0.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter-server==1.2.1
jupyterlab==3.0.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.2
jupyterlab-widgets==1.0.0
macholib==1.14
MarkupSafe==1.1.1
mistune==0.8.4
moderngl==5.6.2
multipledispatch==0.6.0
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.0
nest-asyncio==1.4.3
notebook==6.2.0
numpy==1.19.2
opensimplex==0.3
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pefile==2019.4.18
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.10
ptyprocess==0.7.0
pycparser==2.20
pydub==0.24.1
Pygments==2.7.2
PyInstaller==3.4
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.1
pyrr==0.10.3
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
pywin32-ctypes==0.2.0
pyzmq==21.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.1
Send2Trash==1.5.0
simpleaudio==1.0.4
six==1.15.0
sniffio==1.2.0
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.1.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
zmq==0.0.0
